separate predict method for ml models under the hood
